// frontend/ticket-booking-web/src/app/api/model/event-dto.ts
/**
 * Ticket Booking Platform API - v1
 *
 * Contact: m1musbal@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { EventCategory } from './event-category';
import { TicketCategoryDto } from './ticket-category-dto';
import { EventStatus } from './event-status';


export interface EventDto { 
    id: string;
    title: string;
    description: string;
    category: EventCategory;
    status: EventStatus;
    startsAt: string;
    endsAt: string;
    locationId: string;
    ticketCategories: Array<TicketCategoryDto>;
}
export namespace EventDto {
}